all-code-commit
Payment Form HTML Template

A simple and user-friendly payment form created in HTML for seamless online transactions. This lightweight and responsive form is designed to enhance the payment experience on your website. It includes essential fields for card details, billing information, and a straightforward interface for a smooth user interaction. Feel free to integrate, customize, and improve upon it for your specific project needs.